+++ ctor/src/parse.rs
@@ -40,7 +40,6 @@ macro_rules! __ctor_parse_impl {
         features = (
             anonymous = $anonymous:tt,
             crate_path = $crate_path:tt,
-            dtor = $dtor:tt,
             export_name_prefix = $export_name_prefix:tt,
             link_section = $link_section:tt,
             no_warn_on_missing_unsafe = $no_warn_on_missing_unsafe:tt,