+++ ctor/tests/expand-linux/dtor.rs
@@ -1,6 +1,0 @@
-use ctor::dtor;
-
-#[dtor]
-unsafe fn foo() {
-    println!("foo");
-}